Main code commit.
Commit data_gen/, which contains the Python FCC Auction instance
generator as well as the shell script used to generate the exact
instances used in HW2.

Commit code/, which contains the C++ code for loading data from
the instance files, writing them as an IP, solving via CPLEX, and
outputting the found objective.